Separate error handling into its own file
